backend/pyproject.toml
[tool.poetry]
name = "MLOps Chatbot"
version = "0.1.0"
description = ""
authors = ["ruizguille <guillermo@codeawake.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
pdfminer-six = "^20250506"
tqdm = "^4.67.1"
tiktoken = "^0.9.0"
nltk = "^3.9.1"
pydantic-settings = "^2.9.1"
redis = "^6.2.0"
numpy = "^2.2.6"
pydantic = "^2.11.5"
rich = "^14.0.0"
openai = "^1.82.1"
fastapi = {extras = ["standard"], version = "^0.115.12"}
sse-starlette = "^2.3.6"
sentence-transformers = "^3.3.1"
qdrant-client = "^1.12.0"
ollama = "^0.6.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
load = "app.loader:main"
local = "app.assistants.local_assistant:main"
export = "app.export:main"